protection

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: rarahhal <rarahhal@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/13 17:19:38 by rarahhal          #+#    #+#             */
-/*   Updated: 2021/11/14 19:06:21 by rarahhal         ###   ########.fr       */
+/*   Updated: 2021/11/26 11:55:11 by rarahhal         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
